remove designation cell v8 file
