+++ components/pages/submission-system/program-dashboard/DonorDataSummary/DesignationCell.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022 The Ontario Institute for Cancer Research. All rights reserved
+ * Copyright (c) 2023 The Ontario Institute for Cancer Research. All rights reserved
  *
  * This program and the accompanying materials are made available under the terms of
  * the GNU Affero General Public License v3.0. You should have received a copy of the
@@ -17,7 +17,7 @@
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-import { css, styled, UikitTheme, useTheme } from '@icgc-argo/uikit';
+import { css, styled, TableCellWrapper, UikitTheme, useTheme } from '@icgc-argo/uikit';
 import { DonorSummaryEntry } from 'generated/gql_types';
 import { ProgramDonorSummaryEntryField } from './types';
 
@@ -29,7 +29,6 @@ const DesignationContainer = styled('div')`
   width: 100%;
   height: 100%;
   margin-left: -8px;
-  position: absolute;
 `;
 
 const DesignationEntry = styled('div')`
@@ -40,6 +39,9 @@ const DesignationEntry = styled('div')`
   line-height: 28px;
   flex: 1;
   background: ${({ background }: { background: string }) => background};
+  display: flex;
+  justify-content: center;
+  align-items: center;
 `;
 
 export const DesignationCell = ({
@@ -92,17 +94,22 @@ export const DesignationCell = ({
     : errorColors.default;
 
   return (
-    <DesignationContainer>
+    <TableCellWrapper
+      css={css`
+        padding: 0 !important;
+        align-items: stretch !important;
+      `}
+    >
       <DesignationEntry background={normalBg}>{normalCount}N</DesignationEntry>
       <DesignationEntry background={tumourBg}>{tumourCount}T</DesignationEntry>
-    </DesignationContainer>
+    </TableCellWrapper>
   );
 };
 
 // FEATURE_PROGRAM_DASHBOARD_RNA_ENABLED - remove when flag enabled in production
 // this code isn't required once the RNA feature is in production
 const isValid = (num: number) => num > 0;
-const DesignationEntryStyled = styled('div')`
+const DesignationEntryLegacy = styled('div')`
   text-align: center;
   flex: 1;
   color: ${({ num, theme }: { num: number; theme: UikitTheme }) =>
@@ -112,10 +119,10 @@ export const DesignationCellLegacy = ({ left, right }: { left: number; right: nu
   const theme = useTheme();
   return (
     <DesignationContainer>
-      <DesignationEntryStyled theme={theme} num={left}>
+      <DesignationEntryLegacy theme={theme} num={left}>
         {left}N
-      </DesignationEntryStyled>
-      <DesignationEntryStyled
+      </DesignationEntryLegacy>
+      <DesignationEntryLegacy
         theme={theme}
         num={right}
         css={css`
@@ -123,7 +130,7 @@ export const DesignationCellLegacy = ({ left, right }: { left: number; right: nu
         `}
       >
         {right}T
-      </DesignationEntryStyled>
+      </DesignationEntryLegacy>
     </DesignationContainer>
   );
 };
